Animations: Lancelot test for animations
Add animation test files that test different properties. A frame will be
rendered in the middle of the animation and used for testing. Also,
QSvgTinyDocument is used instead of QSvgRenderer to access the manually
controlled animator. Finally, if the document is not loaded, return a white filled image instead of an image of size 0 that won't be registered in the baseline

Change-Id: Ia8fe6a7f1a6d708df079db13bcd5f0ebccad01b2
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>